// src/views/main/types/accessCtl.ts
export type AccessCtlLog = {
  LogIDX: number | null;
  LogStatus: string | null;
  LogStatusName: string | null;
  LogDoorID: string | null;
  LogDoorName: string | null;
  LogPersonID: string | null;
  LogPersonLastName: string | null;
  LogTitleName: string | null;
  LogDateTime: string | null;
  camera_id?: string | null;
};